requirements.txt
Flask==2.3.2
werkzeug==2.3.6
pdfplumber==0.9.0
reportlab==4.0.4
pytesseract==0.3.10
Pillow==9.5.0
opencv-python==4.8.0.74
transformers==4.30.2
torch==2.2.0
sentencepiece==0.1.99
protobuf==3.20.3
deep-translator==1.11.4
python-multipart==0.0.6
uuid==1.30
python-dotenv==1.0.0
